pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "fortran-legacy-tools"
dynamic = ["version"]
description = "Tools to deal with Fortran code"
dependencies = ["docopt-ng"]

[[project.authors]]
name = "Elias Rabel"

[[project.authors]]
name = "Jacob Williams"

[[project.authors]]
name = "Christopher Saloman"

[project.optional-dependencies]
development = ["wheel", "pytest"]

[project.scripts]
fixed2free = "fortran_legacy_tools.fixed2free2:main"
fdeclarations = "fortran_legacy_tools.fdeclarations:main"
flowercase = "fortran_legacy_tools.flowercase:main"

[tool.setuptools.packages.find]
namespaces = true
where = ["src"]
